pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-server>=0.1.0",
    "google-adk>=0.2.0",
    "litellm>=1.67.2",
    "python-dotenv>=1.1.0",
]
